fix(document enricher):	event dates -> the order of the results is incorrect.
Issue.
The sorting of events with multiple sp_scheduling_iterate is incorrect.  Repeat events that start in the past are always prioritized in order.

Solution.
The idea is to change the list of dates available to the Solr process. Solr only considers the list of current dates for indexing, without dates that are already in the past.